Add Test Runner
I added a simple test runner and hatch testing environment to test the
script's component parsing.

I also added a single argument to the modulr.py script which determines
the project root for the parsing run, but I used sys.argv directly as a
quick and dirty hack. This needs to be updated to use argparse as soon
as possible.