Fix Incorrect Tag Syntax in Generated Stylesheet Links
When I added automatic js and css inclusion, I forgot that `<link>`
os a void tag and can't have a closing tag, so the implementation I
added was creating invalid HTML when including a component-specific
stylesheet.

This commit removes the ending tag from the generated `<link>` and also
updates the test case for it to expect the correct output.