Add GitHub Actions CI Workflow
I added a github actions workflow for running the test suite on PRs and
direct pushes to both dev and main.

For some reason, the bash shell on the windows runner doesn't work when
the shell script is a string in the hatch run command, but it works when
it's executed as a script, so I also added a `test-suite.sh` script that
is run by hatch in order to get the same test suite behavior on Linux,
MacOS, and Windows.

I also added an exit code to the `checkresults.py` script so that failed
tests will properly trigger a failure in the CI.